requirements.txt
streamlit
google-generativeai
matplotlib
networkx
pandas
scikit-learn
numpy